% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{HOTP}
\alias{HOTP}
\title{HMAC based One Time Password (HOTP)}
\description{
An R6 class that implements the HMAC based One Time Password (HOTP) algorithm.
}
\section{Initialization}{
\preformatted{
HOTP$new(secret, digits = 6L, algorithm = "sha1")
}
Create an One Time Password object
\itemize{
\item \strong{secret} a scalar character, the base32-based secret key
\item \strong{digits} an integer, the number of digits of the password
\item \strong{algorithm} the hash algorithm used, possible values are
"sha1", "sha256" and "sha512".
}
}

\section{Methods}{
\preformatted{
HOTP$at(counter)
}
Generate an one time password at counter value
\itemize{
\item \strong{counter} a non-negative integer
}
}

\examples{
p <- HOTP$new("")
}
\seealso{
https://tools.ietf.org/html/rfc4226
}
