% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\docType{package}
\name{otp-package}
\alias{otp}
\alias{otp-package}
\title{otp: One Time Password Generation and Verification}
\description{
Generating and validating One-time Password based on Hash-based Message Authentication Code (HOTP) and Time Based One-time Password (TOTP) according to RFC 4226 \url{https://datatracker.ietf.org/doc/html/rfc4226} and RFC 6238 \url{https://datatracker.ietf.org/doc/html/rfc6238}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randy3k/otp}
  \item Report bugs at \url{https://github.com/randy3k/otp/issues}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com} [copyright holder]

}
